Various small code cleanups: - Lots of "signed vs unsigned" comparison fixes - time_t globally replaced with TIME_T - All seqnos moved to "seqno_t", which defaults to unsigned long - DIMOF() definition centralized to portability.h and typecast to int - EOS define moved to portability.h - dropped inclusion of signal.h from stonith.h, so that sigignore is   properly defined
CVS patchset: 1584
CVS date: 2003/01/31 10:02:09

+++ include/clplumbing/apphb_cs.h
@@ -63,7 +63,7 @@ struct apphb_signupmsg {
 /* App heartbeat setinterval / setwarn message */
 struct apphb_msmsg {
 	char	msgtype [APPHB_TLEN];
-	int	ms;
+	unsigned long	ms;
 };
 
 /* App heartbeat server return code (errno) */
